Stop using PRI*16 and %zu format specifiers (#17502)
* Stop using PRI*16 and %zu format specifiers

Some libcs don't support them, use %u/%d/%x instead

* Commit generated by running the following git-extras commands:
		git sed -f g '" PRIu16 "' 'u'
		git sed -f g '" PRIu16' 'u"'
		git sed -f g '" PRIx16 "' 'x'
		git sed -f g '" PRIx16' 'x"'
		git sed -f g '" PRIX16 "' 'X'
		git sed -f g '" PRIX16' 'X"'
		git sed -f g '" PRId16 "' 'd'
		git sed -f g '" PRId16' 'd"'
		git sed -f g '%zu' '%u'

* Cast some %zu parameters due to warnings
* Add lint rules
* Run restyle


* Change c-style casts to c++-casts and restore variable type changes


* Add parentheses around static_cast arguments


* Run restyle locally

+++ examples/chef/esp32/main/main.cpp
@@ -129,7 +129,7 @@ void DeviceEventCallback(const ChipDeviceEvent * event, intptr_t arg)
         break;
     }
 
-    ChipLogProgress(Shell, "Current free heap: %zu\n", heap_caps_get_free_size(MALLOC_CAP_8BIT));
+    ChipLogProgress(Shell, "Current free heap: %u\n", static_cast<unsigned int>(heap_caps_get_free_size(MALLOC_CAP_8BIT)));
 }
 
 const char * TAG = "chef-app";
